Use `m ~ IO` to improve type inference (#23)
* Use `m ~ IO` to improve type inference

Currently, one has to use the `hedgehog` identity function to force the type of the test to be `PropertyT IO ()`. A better way to do this is to instead use an equality constraint in the instance head. This means that when we typecheck `it` we encounter `PropertyT m ()` as before, but we're now able to find an instance that matches. Refining the context will the force `m` to be `IO` and everything works out fine.

* Update Spec.hs

* Fix the other instance